Purpose:     *Fixed validation errors
Description:
    *Fixed integer conversions for validation routines

+++ branches/fortran_1_8/fortran/test/tH5A_1_8.f90
@@ -3936,7 +3936,6 @@ SUBROUTINE lapl_nlinks( fapl, total_error)
   CALL check("H5Pset_nlinks_f",error,total_error)
   CALL VERIFY("H5Pset_nlinks_f",INT(nlinks), 20, total_error)
 
-
   !/* Open object through what is normally too many soft links using
   ! * new property list */
   
@@ -3968,7 +3967,7 @@ SUBROUTINE lapl_nlinks( fapl, total_error)
 
   CALL H5Pget_nlinks_f(plist, nlinks, error)
   CALL check("H5Pget_nlinks_f",error,total_error)
-  CALL VERIFY("H5Pget_nlinks_f", nlinks, 4, total_error)
+  CALL VERIFY("H5Pget_nlinks_f", INT(nlinks), 4, total_error)
 
   ! /* Try opening through what is now too many soft links */
